pyproject.toml
[project]
name = "trendposter"
version = "0.1.0"
description = "AI-powered tweet scheduler that posts your tweets when they match trending topics"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "tweepy>=4.14",
    "python-dotenv>=1.0",
    "apscheduler>=3.10",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
gemini = ["google-generativeai>=0.8"]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.4"]
all = [
    "anthropic>=0.40",
    "openai>=1.50",
    "google-generativeai>=0.8",
    "python-telegram-bot>=21.0",
    "discord.py>=2.4",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
trendposter = "trendposter.cli:main"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"
